[SYCL-JIT][NFCI] Enable few extra warnings
Those extra flags are required per our internal guidelines (the full
list you can find in `llvm/cmake/modules/AddSecurityFlags.cmake`.

Even though we have this shared cmake module that applies those flags, I
don't think that we will be able to use it for every flag, because
baseline LLVM that we use isn't warning-free. Therefore, those flags
were directly applied to `sycl-jit`.

+++ sycl-jit/jit-compiler/include/DynArray.h
@@ -22,7 +22,7 @@ public:
   explicit DynArray(size_t Size) { init(Size); }
 
   template <typename InputIt> DynArray(InputIt Begin, InputIt End) {
-    init(End - Begin);
+    init(static_cast<size_t>(End - Begin));
     std::copy(Begin, End, this->begin());
   }
